porting fixe non dialog message in ServerTransactionWrapper.
git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/jain-slee@5103 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ resources/sip11/ra/src/main/java/org/mobicents/slee/resource/sip11/wrappers/ServerTransactionWrapper.java
@@ -40,7 +40,7 @@ public class ServerTransactionWrapper extends SuperTransactionWrapper implements
 		int statusCode = arg0.getStatusCode();
 		if (method.equals(Request.CANCEL) && (statusCode < 300 && statusCode > 199)) {
 			Dialog dialog = getDialog();
-			if (dialog.getState() == null)
+			if (dialog!=null && dialog.getState() == null)
 				dialog.delete();
 		}
 	}
